cw5/my_fastapi_project/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
httpx==0.25.0
